+++ example/lib/pages/home_page.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:file_picker/file_picker.dart';
+import 'package:filesystem_picker/filesystem_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -38,7 +39,6 @@ class _HomePageState extends State<HomePage> {
             document: doc, selection: const TextSelection.collapsed(offset: 0));
       });
     } catch (error) {
-      print(error);
       final doc = Document()..insert(0, 'Empty asset');
       setState(() {
         _controller = QuillController(
@@ -99,16 +99,33 @@ class _HomePageState extends State<HomePage> {
         autoFocus: false,
         readOnly: false,
         placeholder: 'Add content',
-        onMentionTap: (val){
-          print("onMentionTap: $val");
-        },
-        onLaunchUrl: (val){
-          print("onLaunchUrl: $val");
-
-        },
         expands: false,
         padding: EdgeInsets.zero,
         customStyles: DefaultStyles(
+          h1: DefaultTextBlockStyle(
+              const TextStyle(
+                fontSize: 32,
+                color: Colors.black,
+                height: 1.15,
+                fontWeight: FontWeight.w300,
+              ),
+              const Tuple2(16, 0),
+              const Tuple2(0, 0),
+              null),
+          sizeSmall: const TextStyle(fontSize: 9),
+        ));
+    if (kIsWeb) {
+      quillEditor = QuillEditor(
+          controller: _controller!,
+          scrollController: ScrollController(),
+          scrollable: true,
+          focusNode: _focusNode,
+          autoFocus: false,
+          readOnly: false,
+          placeholder: 'Add content',
+          expands: false,
+          padding: EdgeInsets.zero,
+          customStyles: DefaultStyles(
             h1: DefaultTextBlockStyle(
                 const TextStyle(
                   fontSize: 32,
@@ -120,44 +137,25 @@ class _HomePageState extends State<HomePage> {
                 const Tuple2(0, 0),
                 null),
             sizeSmall: const TextStyle(fontSize: 9),
-            mentionStyle: const TextStyle(
-                fontSize: 15, color: Colors.red, fontWeight: FontWeight.bold)));
-    if (kIsWeb) {
-      quillEditor = QuillEditor(
-          controller: _controller!,
-          scrollController: ScrollController(),
-          scrollable: true,
-          focusNode: _focusNode,
-          autoFocus: false,
-          readOnly: false,
-          placeholder: 'Add content',
-          expands: false,
-          onMentionTap: (val){
-            print("onMentionTap: $val");
-          },
-          onLaunchUrl: (val){
-            print("onLaunchUrl: $val");
-
-          },
-          padding: EdgeInsets.zero,
-          customStyles: DefaultStyles(
-              h1: DefaultTextBlockStyle(
-                  const TextStyle(
-                    fontSize: 32,
-                    color: Colors.black,
-                    height: 1.15,
-                    fontWeight: FontWeight.w300,
-                  ),
-                  const Tuple2(16, 0),
-                  const Tuple2(0, 0),
-                  null),
-              sizeSmall: const TextStyle(fontSize: 9),
-              mentionStyle: const TextStyle(
-                  fontSize: 15,
-                  color: Colors.red,
-                  fontWeight: FontWeight.bold)),
+          ),
           embedBuilder: defaultEmbedBuilderWeb);
     }
+    var toolbar = QuillToolbar.basic(
+        controller: _controller!, onImagePickCallback: _onImagePickCallback);
+    if (kIsWeb) {
+      toolbar = QuillToolbar.basic(
+          controller: _controller!,
+          onImagePickCallback: _onImagePickCallback,
+          webImagePickImpl: _webImagePickImpl);
+    }
+    final isDesktop = !kIsWeb && !Platform.isAndroid && !Platform.isIOS;
+    if (isDesktop) {
+      toolbar = QuillToolbar.basic(
+          controller: _controller!,
+          onImagePickCallback: _onImagePickCallback,
+          filePickImpl: openFileSystemPickerForDesktop);
+    }
+
     return SafeArea(
       child: Column(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -175,17 +173,20 @@ class _HomePageState extends State<HomePage> {
                   child: Container(
                   padding:
                       const EdgeInsets.symmetric(vertical: 16, horizontal: 8),
-                  child: QuillToolbar.basic(
-                      controller: _controller!,
-                      onImagePickCallback: _onImagePickCallback),
+                  child: toolbar,
                 ))
-              : Container(
-                  child: QuillToolbar.basic(
-                      controller: _controller!,
-                      onImagePickCallback: _onImagePickCallback),
-                ),
+              : Container(child: toolbar)
         ],
       ),
+    );
+  }
+
+  Future<String?> openFileSystemPickerForDesktop(BuildContext context) async {
+    return await FilesystemPicker.open(
+      context: context,
+      rootDirectory: await getApplicationDocumentsDirectory(),
+      fsType: FilesystemType.file,
+      fileTileSelectMode: FileTileSelectMode.wholeTile,
     );
   }
 
@@ -198,6 +199,20 @@ class _HomePageState extends State<HomePage> {
     final copiedFile =
         await file.copy('${appDocDir.path}/${basename(file.path)}');
     return copiedFile.path.toString();
+  }
+
+  Future<String?> _webImagePickImpl(
+      OnImagePickCallback onImagePickCallback) async {
+    final result = await FilePicker.platform.pickFiles();
+    if (result == null) {
+      return null;
+    }
+
+    // Take first, because we don't allow picking multiple files.
+    final fileName = result.files.first.name;
+    final file = File(fileName);
+
+    return onImagePickCallback(file);
   }
 
   Widget _buildMenuBar(BuildContext context) {