+++ lib/src/widgets/editor.dart
@@ -126,6 +126,7 @@ class QuillEditor extends StatefulWidget {
       required this.readOnly,
       required this.expands,
       this.showCursor,
+      this.paintCursorAboveText,
       this.placeholder,
       this.enableInteractiveSelection = true,
       this.scrollBottomInset = 0,
@@ -169,6 +170,7 @@ class QuillEditor extends StatefulWidget {
   final EdgeInsetsGeometry padding;
   final bool autoFocus;
   final bool? showCursor;
+  final bool? paintCursorAboveText;
   final bool readOnly;
   final String? placeholder;
   final bool enableInteractiveSelection;
@@ -298,7 +300,7 @@ class _QuillEditorState extends State<QuillEditor>
             width: 2,
             radius: cursorRadius,
             offset: cursorOffset,
-            paintAboveText: paintCursorAboveText,
+            paintAboveText: widget.paintCursorAboveText ?? paintCursorAboveText,
             opacityAnimates: cursorOpacityAnimates,
           ),
           widget.textCapitalization,