+++ lib/src/widgets/keyboard_listener.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter/widgets.dart';
 
 enum InputShortcut { CUT, COPY, PASTE, SELECT_ALL }
 
@@ -63,16 +63,14 @@ class KeyboardListener {
     LogicalKeyboardKey.keyA: InputShortcut.SELECT_ALL,
   };
 
-  bool handleRawKeyEvent(RawKeyEvent event) {
-    print('keysPressed: ${event.toString()}');
-
+  KeyEventResult handleRawKeyEvent(RawKeyEvent event) {
     if (kIsWeb) {
       // On web platform, we ignore the key because it's already processed.
-      return false;
+      return KeyEventResult.ignored;
     }
 
     if (event is! RawKeyDownEvent) {
-      return false;
+      return KeyEventResult.ignored;
     }
 
     final keysPressed =
@@ -88,7 +86,7 @@ class KeyboardListener {
                 .length >
             1 ||
         keysPressed.difference(_interestingKeys).isNotEmpty) {
-      return false;
+      return KeyEventResult.ignored;
     }
 
     if (_moveKeys.contains(key)) {
@@ -106,6 +104,6 @@ class KeyboardListener {
     } else if (key == LogicalKeyboardKey.backspace) {
       onDelete(false);
     }
-    return false;
+    return KeyEventResult.ignored;
   }
 }