+++ lib/src/widgets/toolbar/image_button.dart
@@ -1,11 +1,8 @@
 import 'dart:io';
 
-import 'package:file_picker/file_picker.dart';
-import 'package:filesystem_picker/filesystem_picker.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:path_provider/path_provider.dart';
 
 import '../../models/documents/nodes/embed.dart';
 import '../controller.dart';
@@ -21,6 +18,8 @@ class ImageButton extends StatelessWidget {
     this.fillColor,
     this.onImagePickCallback,
     this.imagePickImpl,
+    this.filePickImpl,
+    this.webImagePickImpl,
     Key? key,
   }) : super(key: key);
 
@@ -35,7 +34,11 @@ class ImageButton extends StatelessWidget {
 
   final ImagePickImpl? imagePickImpl;
 
+  final WebImagePickImpl? webImagePickImpl;
+
   final ImageSource imageSource;
+
+  final FilePickImpl? filePickImpl;
 
   @override
   Widget build(BuildContext context) {
@@ -47,11 +50,12 @@ class ImageButton extends StatelessWidget {
       hoverElevation: 0,
       size: iconSize * 1.77,
       fillColor: fillColor ?? theme.canvasColor,
-      onPressed: () => _handleImageButtonTap(context),
+      onPressed: () => _handleImageButtonTap(context, filePickImpl),
     );
   }
 
-  Future<void> _handleImageButtonTap(BuildContext context) async {
+  Future<void> _handleImageButtonTap(BuildContext context,
+      [FilePickImpl? filePickImpl]) async {
     final index = controller.selection.baseOffset;
     final length = controller.selection.extentOffset - index;
 
@@ -60,11 +64,17 @@ class ImageButton extends StatelessWidget {
       imageUrl = await imagePickImpl!(imageSource);
     } else {
       if (kIsWeb) {
-        imageUrl = await _pickImageWeb();
+        assert(
+            webImagePickImpl != null,
+            'Please provide webImagePickImpl for Web '
+            '(check out example directory for how to do it)');
+        imageUrl = await webImagePickImpl!(onImagePickCallback!);
       } else if (Platform.isAndroid || Platform.isIOS) {
-        imageUrl = await _pickImage(imageSource);
+        imageUrl = await _pickImage(imageSource, onImagePickCallback!);
       } else {
-        imageUrl = await _pickImageDesktop(context);
+        assert(filePickImpl != null, 'Desktop must provide filePickImpl');
+        imageUrl = await _pickImageDesktop(
+            context, filePickImpl!, onImagePickCallback!);
       }
     }
 
@@ -73,38 +83,24 @@ class ImageButton extends StatelessWidget {
     }
   }
 
-  Future<String?> _pickImageWeb() async {
-    final result = await FilePicker.platform.pickFiles();
-    if (result == null) {
-      return null;
-    }
-
-    // Take first, because we don't allow picking multiple files.
-    final fileName = result.files.first.name;
-    final file = File(fileName);
-
-    return onImagePickCallback!(file);
-  }
-
-  Future<String?> _pickImage(ImageSource source) async {
+  Future<String?> _pickImage(
+      ImageSource source, OnImagePickCallback onImagePickCallback) async {
     final pickedFile = await ImagePicker().getImage(source: source);
     if (pickedFile == null) {
       return null;
     }
 
-    return onImagePickCallback!(File(pickedFile.path));
+    return onImagePickCallback(File(pickedFile.path));
   }
 
-  Future<String?> _pickImageDesktop(BuildContext context) async {
-    final filePath = await FilesystemPicker.open(
-      context: context,
-      rootDirectory: await getApplicationDocumentsDirectory(),
-      fsType: FilesystemType.file,
-      fileTileSelectMode: FileTileSelectMode.wholeTile,
-    );
+  Future<String?> _pickImageDesktop(
+      BuildContext context,
+      FilePickImpl filePickImpl,
+      OnImagePickCallback onImagePickCallback) async {
+    final filePath = await filePickImpl(context);
     if (filePath == null || filePath.isEmpty) return null;
 
     final file = File(filePath);
-    return onImagePickCallback!(file);
+    return onImagePickCallback(file);
   }
 }