+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.6.0]
+* Support Multi Row Toolbar.
+
+## [1.5.0]
+* Remove file_picker dependency.
+
+## [1.4.1]
+* Remove filesystem_picker dependency.
+
+## [1.4.0]
+* Remove path_provider dependency.
+
+## [1.3.4]
+* Add option to paintCursorAboveText.
+
 ## [1.3.3]
 * Upgrade file_picker version.